test-web: Move test-variant handling code into its own files
No behaviour changes.
